feat(cloud-sql/mysql): update to V2 sample
