+++ cloud-sql/mysql/app.yaml
@@ -9,7 +9,7 @@ env_variables:
     DB_USER: <your-username>
     DB_PASS: <your-password>
     DB_NAME: <your-database-name>
-    DB_HOST: cloudsql
+    INSTANCE_HOST: cloudsql
 
 beta_settings:
     cloud_sql_instances: "<project-name>:<region>:<instance-name>=tcp:3306"
